+++ CHANGELOG.md
@@ -5,9 +5,10 @@ Focusses on making what benchee print out configurable to make it fit to your pr
 ## Features (User Facing)
 * The configuration now has a `:print` key where it is possible to configure in a map what benchee prints out during benchmarking. All options are enabled by default (true). Options are:
   * `:benchmarking`  - print when Benchee starts benchmarking a new job (Benchmarking name ..)
-  * `:comparison` - the comparison of the different benchmarking jobs (x times slower than)
   * `:configuration` - a summary of configured benchmarking options including estimated total run time is printed before benchmarking starts
   * `:fast_warning` - warnings are displayed if functions are executed too    fast leading to inaccurate measures
+* There is also a new configuration option for the built-in console formatter. Which is also enabled by default:
+  * `:comparison` - if the comparison of the different benchmarking jobs (x times slower than) is shown
 * The pre-benchmarking output of the configuration now also prints the currently used Erlang and Elixir versions (similar to `elixir -v`)
 * Add a space between the benchmarked time and the unit